perf(lsp)!: ltex_ls_plus has to be turned on manually
ltex_ls_plus is using a lot of ressources and draining my battery. I turned the vim.lsp.config and enable function from automatically executing to a keymap. The lsp can be turned on with <leader>ltx to attach to all buffers. This way I can work on a document and turn on grammar lsp on command